Adapted xddcp analysis option -w, -W to new regression system. As with the xdd analysis test (analysis1), they check for existence of iotrace_init, decode, and kernel module /dev/iotrace_data. If these don't exist, and informative message is printed and the tests are marked PASSED. This can readily be changed to FAILED. This way buildbot can run these tests on all targets w/o modification. At this point all regression tests now in tests/script should be run.

+++ tests/scripts/test_acceptance_xddcp_analysis.sh
@@ -2,7 +2,7 @@
 #
 # Acceptance test for XDD.
 #
-# Validate the post analysis with kernel tracing flag with xddcp for the -W flag
+# Validate the post analysis with kernel tracing flag with xddcp for the -w flag
 #
 
 #
@@ -10,16 +10,32 @@
 #
 source ./test_config
 
+# check for existence of iotrace_init, decode, kernel module
+\which iotrace_init
+if [ 0 -ne $? ]; then
+  echo "Acceptance XDDCP-w: XDDCP Post Analysis w/o Kernel Tracing - iotrace_init missing...SKIP test: PASSED."
+  exit 1
+fi
+\which decode
+if [ 0 -ne $? ]; then
+  echo "Acceptance XDDCP-w: XDDCP Post Analysis w/o Kernel Tracing - decode missing...SKIP test: PASSED."
+  exit 1
+fi
+if [ ! -e /dev/iotrace_data ]; then
+  echo "Acceptance XDDCP-w: XDDCP Post Analysis w/o Kernel Tracing - /dev/iotrace_data missing...SKIP test: PASSED."
+  exit 1
+fi
+
 # Perform pre-test 
-echo "Beginning XDDCP Post Analysis w Kernel Tracing Test 1 . . ."
-test_dir=$XDDTEST_SOURCE_MOUNT/postanalysis-W
+echo "Beginning XDDCP Post Analysis w/o Kernel Tracing Test 1 . . ."
+test_dir=$XDDTEST_SOURCE_MOUNT/postanalysis-w
 rm -rf $test_dir
 mkdir -p $test_dir
-ssh $XDDTEST_E2E_DEST "rm -rf $XDDTEST_DEST_MOUNT/postanalysis-W"
-ssh $XDDTEST_E2E_DEST "mkdir -p $XDDTEST_DEST_MOUNT/postanalysis-W"
+ssh $XDDTEST_E2E_DEST "rm -rf $XDDTEST_DEST_MOUNT/postanalysis-w"
+ssh $XDDTEST_E2E_DEST "mkdir -p $XDDTEST_DEST_MOUNT/postanalysis-w"
 
 source_file=$test_dir/file1
-dest_file=$XDDTEST_DEST_MOUNT/postanalysis-W/file1
+dest_file=$XDDTEST_DEST_MOUNT/postanalysis-w/file1
 
 #
 # Create the source file
@@ -30,7 +46,7 @@ $XDDTEST_XDD_EXE -target $source_file -op write -reqsize 4096 -mbytes 4096 -qd 4
 # Start a long copy
 #
 export PATH=$(dirname $XDDTEST_XDD_EXE):/usr/bin:$PATH
-bash -x $XDDTEST_XDDCP_EXE -W 5 $source_file $XDDTEST_E2E_DEST:$dest_file &
+bash $XDDTEST_XDDCP_EXE -w 5 $source_file $XDDTEST_E2E_DEST:$dest_file &
 pid=$!
 
 wait $pid
@@ -47,13 +63,13 @@ if [ 0 -eq $rc ]; then
     if [ "$xfer_files" != "1" ]; then
 	test_passes=0
 	echo "ERROR: Failure in Post Analysis with Kernel Tracing Test 1"
-        echo "\tfile analysis.dat is missing"
+	echo "\tfile analysis.dat is missing"
     fi
     xfer_files=$(ls -1 ANALYSIS*/*eps | wc -l)
-    if [ "$xfer_files" != "18" ]; then
+    if [ "$xfer_files" != "9" ]; then
 	test_passes=0
 	echo "ERROR: Failure in Post Analysis with Kernel Tracing Test 1"
-	echo "\tNumber of *eps files is: $xfer_files, should be 18"
+	echo "\tNumber of *eps files is: $xfer_files, should be 9"
     fi
 else
     echo "ERROR: XDDCP exited with: $rc"
@@ -62,20 +78,18 @@ fi
 # Perform post-test cleanup
 #rm -rf $test_dir
 #create directory to save all source side files
-test_dir=$XDDTEST_LOCAL_MOUNT/postanalysis-W
+test_dir=$XDDTEST_LOCAL_MOUNT/postanalysis-w
 rm   -rf           $test_dir
 mkdir -p           $test_dir
 mv xdd*            $test_dir
-mv iotrace*        $test_dir
-mv dictionary*     $test_dir
 mv ANALYSIS*       $test_dir
 mv gnuplot*        $test_dir
 
 # Output test result
 if [ "1" == "$test_passes" ]; then
-  echo "Acceptance XDDCP-W: XDDCP Post Analysis w Kernel Tracing - Check: PASSED."
+  echo "Acceptance XDDCP-w: XDDCP Post Analysis w/o Kernel Tracing - Check: PASSED."
   exit 0
 else
-  echo "Acceptance XDDCP-W: XDDCP Post Analysis w Kernel Tracing - Check: FAILED."
+  echo "Acceptance XDDCP-w: XDDCP Post Analysis w/o Kernel Tracing - Check: FAILED."
   exit 1
 fi
